requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
tqdm==4.66.1
requests==2.31.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.36